==> pegasus_global_analyses/1_eda_assembly.Rmd <==
---
title: "PEGASUS Metabolomics - Assembly & Exploratory Data Analysis"
author: "MHuang"
date: "5/20/2021"
output: 
  html_document: 
    toc: yes
    toc_float: 
      collapsed: false
      smooth_scroll: false
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include = FALSE}

## Packages

pkg <- c("tidyverse", "data.table", "here", "readxl", "e1071", "table1")
for (p in pkg) {
        if (require(p, character.only = T)) {
                print(paste0(p, " loaded successfully"))
        } else {
                install.packages(p)
                require(p, character.only = T)
                print(paste0(p, " downloaded and loaded successfully"))
        }
}


## Paths

pegasus_dir <- "C:/Users/MHuang/Dropbox (Partners HealthCare)/PEGASUS/"
met_dir <- str_c(pegasus_dir, "global_metabolomics_082020/pegasus_global_data_qc/post_qc_data/")
pheno_dir <- str_c(pegasus_dir, "rpdr_data/rpdr_data_wrangling/data/")

dat_dir <- "data/"


## Filenames

all_subj_fname <- str_c(pheno_dir, "all_subj_enc_bmi.csv")
med_cs_fname <- str_c(pheno_dir, "med_corticosteroids_summary.csv")
wbc_fname <- str_c(pheno_dir, "wbc_count_closest_to_serum_collection.csv")
exac_fname <- str_c(pheno_dir, "asthma_exacerbation_summary.csv")
comorb_fname <- str_c(pheno_dir, "comorbidities_of_interest.csv")
pft_fname <- str_c(pheno_dir, "pft_prebronchodilator_closest_to_serum_collection.csv")
ige_fname <- str_c(pheno_dir, "ige_closest_to_serum_collection.csv")
allergy_fname <- str_c(pheno_dir, "ige_high_env_allergy_ever.csv")

hilic_pos_fname <- str_c(met_dir, "hilic_pos/outlier_checked/processed_hilic_pos_halfmin.RData")
lipid_pos_fname <- str_c(met_dir, "lipid_pos/processed_lipid_pos_halfmin.RData")
lipid_neg_fname <- str_c(met_dir, "lipid_neg/processed_lipid_neg_halfmin.RData")
sph_fname <- str_c(pegasus_dir, "sphingolipids_targeted/data/eda_updated.RData")
microb_mets_fname <- str_c(pegasus_dir, "precion_targeted/data/PR003-21_AsthmaCohort1125_MicrobiomePanel_CDT_Test.xlsx")
link_id_fname <- str_c(pegasus_dir, "precion_targeted/data/list_of_BIOBANK_samples.csv")


## Hard-coded numbers

missing_thld <- 0.30

```

# 1. Phenotype/EMR data

```{r}

all_subj <- fread(all_subj_fname)
all_subj[, Collect_Date := as.Date(Collect_Date, format = "%Y-%m-%d")]
all_subj[, .N, .(is.na(EMPI), is.na(Collect_Date))][order(is.na, is.na.1)] %>% 
        setnames(., c("missing_EMPI", "missing_Collect_Date", "N")) %>% print()
all_subj[, asthma_num := case_when(Asthma_Status == TRUE ~ 1, 
                                   Asthma_Status == FALSE ~ 0)]

stratum_to_rm <- all_subj[is.na(EMPI) | is.na(Collect_Date), unique(Stratum)]; length(stratum_to_rm)
all_subj <- all_subj[!(Stratum %in% stratum_to_rm), ]

med_cs <- fread(med_cs_fname)
pheno <- merge(all_subj, med_cs, by = c("EMPI", "SampleGroup"), all.x = T)

comorb <- fread(comorb_fname)
pheno <- merge(pheno, comorb[!is.na(EMPI), ], by = c("EMPI"), all.x = T)

wbc <- fread(wbc_fname)
pheno <- merge(pheno, wbc, by = c("EMPI"), all.x = T)

exac <- fread(exac_fname)
pheno <- merge(pheno, exac, by = c("EMPI", "SampleGroup"), all.x = T)

pft <- fread(pft_fname)
pheno <- merge(pheno, pft, by = c("EMPI", "SampleGroup"), all.x = T)

ige <- fread(ige_fname)
pheno <- merge(pheno, ige, by = c("EMPI"), all.x = T)

allergy <- fread(allergy_fname)
pheno <- merge(pheno, allergy, by = c("EMPI"), all.x = T)

pheno[, ':='(race = case_when(Race %in% c("White", "Black") ~ Race, 
                              TRUE ~ "Other"), 
             fev1fvc_prebd_pctpred = fev1_prebd_pctpred / fvc_prebd_pctpred)]

save(pheno, file = here(dat_dir, "pheno.RData"))

```

# 2. HILIC-positive halfmin imputed data (outlier checked)

```{r}

load(hilic_pos_fname, verbose = T)
hilic_pos_dat <- copy(samp_dat_final)
hilic_pos_mets_info <- copy(mets_info)
hilic_pos_mets_cols <- copy(samp_mets_cols)

hilic_pos_mets_info[, .N, pct_na_met <= missing_thld]


## Create another unique feature identifier to avoid possible repeated IDs with other platform

setnames(hilic_pos_dat, hilic_pos_mets_cols, str_c(hilic_pos_mets_cols, "_hilic_pos"))
hilic_pos_mets_info[, feat_id := str_c(feature, "_hilic_pos")]
hilic_pos_mets_cols <- str_c(hilic_pos_mets_cols, "_hilic_pos")


## Manually annotate cortisol (F1381, confirmed with standard by Pei)

hilic_pos_mets_info[feature == "F1381", tentative_annotation := "Cortisol"]


## Add indicators for platform, medication & without MS2

hilic_pos_mets_info[, platform := "HILIC-pos"]

meds_string <- str_c("Flutrimazole|Bezafibrate|Esomeprazole|Levocetirizine|Paracetamol|Irbesartan|",
                     "Gabapentin|Fexofenadine|Isosorbide|Lisofylline|Venlafaxine|Nortriptyline|Metformin|", 
                     "Escitalopram|Methocarbamol|Dantrolene|Trimethoprim|Sulfamethoxazole|Thioguanine|", 
                     "Thalidomide|Diazepam|Trazodone|Levetiracetam|Diltiazem|Fenoterol|Eflornithine|", 
                     "Carbetapentane|Fluoxetine|Fluconazole|Metronidazole|Valsartan|Warfarin")

hilic_pos_mets_info[tentative_annotation != "Unknown", 
                    ':='(wo_ms2 = case_when(grepl("(?i)w/o ms2", tentative_annotation) ~ 1, 
                                            TRUE ~ 0), 
                         med_feat = case_when(grepl(meds_string, tentative_annotation, ignore.case = T) ~ 1, 
                                              TRUE ~ 0))]


## Remove original metabolomic data files with non-specific names

rm(mets_info, samp_dat_final, samp_dat_origscale, samp_na, samp_mets_cols)
save(hilic_pos_dat, hilic_pos_mets_info, hilic_pos_mets_cols, 
     file = here(dat_dir, "hilic_pos_data_updated.RData"))

```

# 3. Lipidomics-positive halfmin imputed data 

```{r}

load(lipid_pos_fname, verbose = T)
setnames(samp_dat_final, "Group", "SampleGroup")
lipid_pos_dat <- copy(samp_dat_final)
lipid_pos_mets_info <- copy(mets_info)
lipid_pos_mets_cols <- copy(samp_mets_cols)

lipid_pos_mets_info[, .N, pct_na_met <= missing_thld]


## Create another unique feature identifier to avoid possible repeated IDs with other platform

setnames(lipid_pos_dat, lipid_pos_mets_cols, str_c(lipid_pos_mets_cols, "_lipid_pos"))
lipid_pos_mets_info[, feat_id := str_c(feature, "_lipid_pos")]
lipid_pos_mets_cols <- str_c(lipid_pos_mets_cols, "_lipid_pos")


## Manually annotate mouse/cell feature to unknown

lipid_pos_mets_info[grepl("(?i)mouse|cell", tentative_annotation), tentative_annotation := "Unknown"]


## Add indicators for platform & without MS2

lipid_pos_mets_info[, platform := "Lipid-pos"]

lipid_pos_mets_info[tentative_annotation != "Unknown", 
                    ':='(wo_ms2 = case_when(grepl("(?i)w/o ms2", tentative_annotation) ~ 1, 
                                            TRUE ~ 0), 
                         med_feat = 0)] 


## Remove original metabolomic data files with non-specific names

rm(mets_info, samp_dat_final, samp_dat_origscale, samp_na, samp_mets_cols)
save(lipid_pos_dat, lipid_pos_mets_info, lipid_pos_mets_cols, 
     file = here(dat_dir, "lipid_pos_data_updated.RData"))

```

# 4. Lipidomics-negative halfmin imputed data 

```{r}

load(lipid_neg_fname, verbose = T)
lipid_neg_dat <- copy(samp_dat_final)
lipid_neg_mets_info <- copy(mets_info)
lipid_neg_mets_cols <- copy(samp_mets_cols)

lipid_neg_mets_info[, .N, pct_na_met <= missing_thld]


## Create another unique feature identifier to avoid possible repeated IDs with other platform

setnames(lipid_neg_dat, lipid_neg_mets_cols, str_c(lipid_neg_mets_cols, "_lipid_neg"))
lipid_neg_mets_info[, feat_id := str_c(feature, "_lipid_neg")]
lipid_neg_mets_cols <- str_c(lipid_neg_mets_cols, "_lipid_neg")


## Manually annotate mouse/cell feature to unknown

lipid_neg_mets_info[grepl("(?i)mouse|cell", tentative_annotation), tentative_annotation := "Unknown"]


## Add indicators for platform & without MS2

lipid_neg_mets_info[, platform := "Lipid-neg"]

lipid_neg_mets_info[tentative_annotation != "Unknown", 
                    ':='(wo_ms2 = case_when(grepl("(?i)w/o ms2", tentative_annotation) ~ 1, 
                                            TRUE ~ 0), 
                         med_feat = 0)] 


## Remove original metabolomic data files with non-specific names

rm(mets_info, samp_dat_final, samp_dat_origscale, samp_na, samp_mets_cols)
save(lipid_neg_dat, lipid_neg_mets_info, lipid_neg_mets_cols, 
     file = here(dat_dir, "lipid_neg_data_updated.RData"))

```

# 5. Targeted sphingolipids data

```{r}

load(sph_fname, verbose = T)
setnames(mets_pheno, c("Group", "Original_Spl_ID"), c("SampleGroup", "sampleID_biobank"))
sph_dat <- copy(mets_pheno)
sph_mets_info <- copy(mets_info)
sph_mets_cols <- copy(sph_list)

sph_mets_info[, .N, pct_na_met <= missing_thld]
sph_mets_info[, summary(pct_na_met)]


## Standardize sphingolipids

sph_dat_orig <- copy(sph_dat)
sph_dat[, (sph_mets_cols) := lapply(.SD, function(x) scale(x, center = T, scale = T)), 
        .SDcols = sph_mets_cols]


## Add indicators for platform & without MS2

sph_mets_info[, ':='(feature = metabolite, 
                     tentative_annotation = metabolite, 
                     platform = "Targeted-sphingolipids", 
                     wo_ms2 = 0, 
                     med_feat = 0, 
                     rt_min = NA, 
                     mz = NA, 
                     cv = NA, 
                     skewness_post = skewness_met)][, feat_id := feature] 
# The annotations are accurate - only duplicating for programing purpose


## Remove original metabolomic data files with non-specific names

rm(mets_pheno, mets_info, sph_list, sphoth_list, sm_list, cer_list, ceroth_list)
save(sph_dat, sph_mets_info, sph_mets_cols, sph_dat_orig, 
     file = here(dat_dir, "sph_data_updated.RData"))

```

# 6. Targeted microbial metabolites data (Precion)

```{r}

link_id <- fread(link_id_fname)
setnames(link_id, "SUBJ_ALIASID", "Biobank_Subject_ID")
# link_id[, .N, Biobank_Subject_ID][N == 2] 
# Remvoe duplicated records for this subject
link_id <- link_id[!duplicated(Biobank_Subject_ID), ]

microb_mets <- read_xlsx(microb_mets_fname) %>% as.data.table()
setnames(microb_mets, gsub("[ ()/]", "_", colnames(microb_mets)))
microb_mets[, feat_id := str_c("X", seq(1, .N), "_microb")]
setcolorder(microb_mets, "feat_id")
microb_mets_info <- microb_mets[, c(1:which(colnames(microb_mets) == "ALOQ__ug_mL_")), with = F]
microb_mets_cols <- microb_mets_info$feat_id

subj_cols <- colnames(microb_mets)[-c(2:which(colnames(microb_mets) == "ALOQ__ug_mL_"))]
microb_mets_dat <- microb_mets[, ..subj_cols] %>% as.data.frame() 
rownames(microb_mets_dat) <- microb_mets_dat$feat_id
microb_mets_dat$feat_id <- NULL
microb_mets_dat <- t(microb_mets_dat) %>% as.data.frame() %>% rownames_to_column() %>% as.data.table()
setnames(microb_mets_dat, "rowname", "S_SAMPLEID")

microb_mets_dat[, (microb_mets_cols) := lapply(.SD, as.character), .SDcols = microb_mets_cols]
suppressWarnings(microb_mets_dat[, (microb_mets_cols) := lapply(.SD, as.numeric), .SDcols = microb_mets_cols])
microb_mets_dat <- merge(microb_mets_dat, link_id[, .(Biobank_Subject_ID, S_SAMPLEID)], by = "S_SAMPLEID")

microb_mets_info[, .N, .(SUB_PATHWAY, SUPER_PATHWAY)]
microb_mets_info[SUB_PATHWAY == "polyamine metabolism", SUB_PATHWAY := "Polyamine Metabolism"]


## Create a variable for tryptophan/kynurenine ratio

microb_mets_dat[, ':='(trypkynu_ratio_microb = X66_microb / X41_microb)]
ggplot(microb_mets_dat) + 
        geom_histogram(aes(trypkynu_ratio_microb), bins = 100) + 
        labs(title = "Microbial metabolites - tryptophan/kynurenine ratio distribution")
microb_mets_cols <- c(microb_mets_cols, "trypkynu_ratio_microb")
microb_mets_info <- rbind(microb_mets_info, 
                          t(c("trypkynu_ratio_microb", "tryptophan/kynurenine ratio", "microbiome metabolism", 
                              "Tryptophan Metabolism", NA, NA, NA, NA, NA, NA, NA)), use.names = F)


## Missingness

mets_missing <- microb_mets_dat[, lapply(.SD, function(x) mean(is.na(x))), .SDcols = microb_mets_cols] %>% 
        t() %>% as.data.frame() %>% rownames_to_column()
setnames(mets_missing, c("feat_id", "pct_na_met"))
microb_mets_info <- merge(microb_mets_info, mets_missing, by = "feat_id")
microb_mets_info[, summary(pct_na_met)]
microb_mets_info[, .N, pct_na_met <= missing_thld]
ggplot(microb_mets_info) + 
        geom_histogram(aes(pct_na_met), bins = 50) + 
        labs(title = "Microbial metabolites missingness")


## Skewness

mets_skewness <- microb_mets_dat[, lapply(.SD, function(x) skewness(is.na(x))), .SDcols = microb_mets_cols] %>% 
        t() %>% as.data.frame() %>% rownames_to_column()
setnames(mets_skewness, c("feat_id", "skewness_met"))
microb_mets_info <- merge(microb_mets_info, mets_skewness, by = "feat_id")
microb_mets_info[, summary(skewness_met)]
ggplot(microb_mets_info) + 
        geom_histogram(aes(skewness_met), bins = 50) + 
        labs(title = "Microbial metabolites skewness")


## Standardize microbial metabolites

microb_mets_dat_orig <- copy(microb_mets_dat)
microb_mets_dat[, (microb_mets_cols) := lapply(.SD, function(x) scale(x, center = T, scale = T)), 
                .SDcols = microb_mets_cols]


## Add indicators for platform & without MS2

microb_mets_info[, ':='(feature = feat_id, 
                        tentative_annotation = Component_Name, 
                        platform = "Targeted-microbial-mets", 
                        wo_ms2 = 0, 
                        med_feat = 0, 
                        rt_min = NA, 
                        mz = NA, 
                        cv = NA, 
                        skewness_post = skewness_met)] 
# The annotations are accurate - only duplicating for programing purpose

save(microb_mets_dat, microb_mets_info, microb_mets_cols, microb_mets_dat_orig, 
     file = here(dat_dir, "microb_data_updated.RData"))

```

# 7. Merge data from all platforms

```{r}

## Merge pheno with metabolomic data

rm_cols <- c("Order", "Batch", "RunID", "SampleType", "QC", "Idx", "BiobankSubjectID", "SampleID", 
             "HILICPOS_dataFileName", str_c("PC", 1:10))
hilic_pos_dat <- hilic_pos_dat[, (rm_cols) := NULL]
pheno_mets <- merge(pheno, hilic_pos_dat, 
                    by = c("sampleID_biobank", "SampleGroup", "Match_ID", "Asthma_Status", 
                           "Age", "Gender", "Race", "BMI", "Nonsmoker"), 
                    all.x = T)

rm_cols <- c("Order", "Batch", "RunID", "SampleType", "QC", "Idx", "BiobankSubjectID", "SampleID", 
             "FileName", str_c("PC", 1:10))
lipid_pos_dat <- lipid_pos_dat[, (rm_cols) := NULL]
pheno_mets <- merge(pheno_mets, lipid_pos_dat, 
                    by = c("sampleID_biobank", "SampleGroup", "Match_ID", "Asthma_Status", 
                           "Age", "Gender", "Race", "BMI", "Nonsmoker"), 
                    all.x = T)

lipid_neg_dat <- lipid_neg_dat[, (rm_cols) := NULL]
pheno_mets <- merge(pheno_mets, lipid_neg_dat, 
                    by = c("sampleID_biobank", "SampleGroup", "Match_ID", "Asthma_Status", 
                           "Age", "Gender", "Race", "BMI", "Nonsmoker"), 
                    all.x = T)

pheno_mets <- merge(pheno_mets, subset(sph_dat, select = c("sampleID_biobank", sph_mets_cols)), 
                    by = "sampleID_biobank", all.x = T)

pheno_mets <- merge(pheno_mets, microb_mets_dat, by = c("Biobank_Subject_ID"), all.x = T)


## Concatenate mets info

mets_info_cols <- c("feat_id", "feature", "tentative_annotation", "platform", "wo_ms2", "med_feat", 
                    "rt_min", "mz", "cv", "pct_na_met", "skewness_post")
mets_info_all <- rbind(hilic_pos_mets_info[, ..mets_info_cols], lipid_pos_mets_info[, ..mets_info_cols], 
                       lipid_neg_mets_info[, ..mets_info_cols], sph_mets_info[, ..mets_info_cols], 
                       microb_mets_info[, ..mets_info_cols])
nrow(mets_info_all) == length(unique(mets_info_all$feat_id)) 


## Generate list of metabolite column names for later analysis

mets_list_analysis <- c(hilic_pos_mets_info[is.na(qc_rm) & pct_na_met <= missing_thld, feat_id], 
                        lipid_pos_mets_info[is.na(qc_rm) & pct_na_met <= missing_thld, feat_id], 
                        lipid_neg_mets_info[is.na(qc_rm) & pct_na_met <= missing_thld, feat_id], 
                        sph_mets_info[pct_na_met <= missing_thld, feat_id], 
                        microb_mets_info[pct_na_met <= missing_thld, feat_id])
length(mets_list_analysis)

```

# 8. EDA
## 8.1. Apply exclusion

 - _trim: subjects missing collection date / bmi / bmi >5yrs away from collection date & matched subjects removed  
 - _excl: further excluded subjects with the following comorbidities and matched subjects  
     - COPD, A1AD, cystic fibrosis, bronchiectasis, chronic bronchitis, immuno deficiency

```{r}

nrow(pheno_mets)

stratum_to_rm <- pheno_mets[is.na(bmi) | abs(bmi_timegap_y) > 5, unique(Stratum)]
stratum_to_rm <- c(stratum_to_rm, pheno_mets[, .N, Stratum][N == 1, Stratum])
print(str_c("Number of strata removed due to missing BMI or BMI outside +/- 5 years of serum date: ", 
            length(stratum_to_rm)))
pheno_mets_trim <- pheno_mets[!(Stratum %in% stratum_to_rm), ]

pheno_mets_trim[, .N, .(ocs_ard_serum_coll, cortisone_ard_serum_coll, SampleGroup, Asthma_Status)
                ][order(Asthma_Status, ocs_ard_serum_coll)]
pheno_mets_trim[, .N, .(copd_5, a1ad, cf, bronchiectasis, cbronchitis, immunodef, Asthma_Status)
                ][order(Asthma_Status)]
comorb_excl <- pheno_mets_trim[copd_5 == 1 | a1ad == 1 | cf == 1 | bronchiectasis == 1 | cbronchitis == 1 | 
                                 immunodef == 1, unique(Stratum)]
print(str_c("Number of strata removed due to comorbidities: ", length(comorb_excl)))
pheno_mets_excl <- pheno_mets_trim[!(Stratum %in% comorb_excl), ]

```

## 8.2. Table 1
### 8.2.1. trim dataset

```{r}

table1( ~ Age + bmi + Gender + Race + Ethnicity + Nonsmoker | SampleGroup, data = pheno_mets_trim)
table1( ~ ics_trim_totnum_1y + ocs_trim_totnum_1y + eos + neut + asthma_exacerbation_trim_totnum_1y + 
          fev1_prebd_pctpred + fvc_prebd_pctpred + log10(ige) | SampleGroup, data = pheno_mets_trim)

```

### 8.2.2. excl dataset

Use this for Table 1

```{r}

table1( ~ Age + bmi + Gender + Race + Ethnicity + Nonsmoker | SampleGroup, data = pheno_mets_excl)
table1( ~ ics_trim_totnum_1y + ocs_trim_totnum_1y + eos + neut + asthma_exacerbation_trim_totnum_1y + 
          fev1_prebd_pctpred + fvc_prebd_pctpred + log10(ige) | SampleGroup, data = pheno_mets_excl)

```

## 8.3. Effective number of test (ENT)
### 8.3.1. In all subjects with serum collection date & RPDR data

```{r}

mets_list_global <- c(hilic_pos_mets_info[is.na(qc_rm) & pct_na_met <= missing_thld, feat_id], 
                      lipid_pos_mets_info[is.na(qc_rm) & pct_na_met <= missing_thld, feat_id], 
                      lipid_neg_mets_info[is.na(qc_rm) & pct_na_met <= missing_thld, feat_id])
pca <- prcomp(pheno_mets[, ..mets_list_global])

print(str_c("Number of PCs needed to account for 80% of all variance in global metabolome: ", 
            which(cumsum(pca$sdev*2)/sum(pca$sdev^2) >= 0.8)[1]))
print(str_c("ENT80% + bonferroni corrected -log10(p-value): ", 
            round(-log10(0.05 / which(cumsum(pca$sdev*2)/sum(pca$sdev^2) >= 0.8)[1]), 3)))

print(str_c("Number of PCs needed to account for 90% of all variance in global metabolome: ", 
            which(cumsum(pca$sdev*2)/sum(pca$sdev^2) >= 0.9)[1]))
print(str_c("ENT90% + bonferroni corrected -log10(p-value): ", 
            round(-log10(0.05 / which(cumsum(pca$sdev*2)/sum(pca$sdev^2) >= 0.9)[1]), 3)))

```

### 8.3.2. In subjects with BMI within +/- 5 years of serum date (trim dataset)

```{r}

pca_trim <- prcomp(pheno_mets_trim[, ..mets_list_global])

print(str_c("Number of PCs needed to account for 80% of all variance in global metabolome: ", 
            which(cumsum(pca_trim$sdev*2)/sum(pca_trim$sdev^2) >= 0.8)[1]))
print(str_c("ENT80% + bonferroni corrected -log10(p-value): ", 
            round(-log10(0.05 / which(cumsum(pca_trim$sdev*2)/sum(pca_trim$sdev^2) >= 0.8)[1]), 3)))

print(str_c("Number of PCs needed to account for 90% of all variance in global metabolome: ", 
            which(cumsum(pca_trim$sdev*2)/sum(pca_trim$sdev^2) >= 0.9)[1]))
print(str_c("ENT90% + bonferroni corrected -log10(p-value): ", 
            round(-log10(0.05 / which(cumsum(pca_trim$sdev*2)/sum(pca_trim$sdev^2) >= 0.9)[1]), 3)))

```

### 8.3.3. In subjects with BMI within +/- 5 years & comorbidities excluded (excl dataset)

Variation accounted for here does not amount to 90% due to n<p; highest is 88.3%

```{r}

pca_excl <- prcomp(pheno_mets_excl[, ..mets_list_global])

print(str_c("Number of PCs needed to account for 80% of all variance in global metabolome: ", 
            which(cumsum(pca_excl$sdev*2)/sum(pca_excl$sdev^2) >= 0.8)[1]))
print(str_c("ENT80% + bonferroni corrected -log10(p-value): ", 
            round(-log10(0.05 / which(cumsum(pca_excl$sdev*2)/sum(pca_excl$sdev^2) >= 0.8)[1]), 3)))

print(str_c("Number of PCs needed to account for 90% of all variance in global metabolome: ", 
            which(cumsum(pca_excl$sdev*2)/sum(pca_excl$sdev^2) >= 0.85)[1]))
print(str_c("ENT85% + bonferroni corrected -log10(p-value): ", 
            round(-log10(0.05 / which(cumsum(pca_excl$sdev*2)/sum(pca_excl$sdev^2) >= 0.85)[1]), 3)))

```

## 8.4. Save merge data

```{r}

save(pheno_mets, pheno_mets_trim, pheno_mets_excl, mets_info_all, mets_list_analysis, 
     hilic_pos_mets_info, lipid_pos_mets_info, lipid_neg_mets_info, sph_mets_info, microb_mets_info, 
     file = here(dat_dir, "all_platforms_data_updated.RData"))

```

# Session info

```{r}

timestamp()
sessionInfo()

```
