==> pegasus_global_analyses/3_signif_features_figures.Rmd <==
---
title: "PEGASUS Metabolomics - Significant Features & Figures"
author: "MHuang"
date: "5/24/2020"
output: 
  html_document: 
    toc: yes
    toc_float:
      collapsed: false
      smooth_scroll: false
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include = FALSE}

## Packages

pkg <- c("tidyverse", "data.table", "here", "readxl", "gtools", "DT", "corrplot", "pheatmap", "RColorBrewer", 
         "viridis", "rstatix", "ggvenn", "ggrepel", "scales")
for (p in pkg) {
        if (require(p, character.only = T)) {
                print(paste0(p, " loaded successfully"))
        } else {
                install.packages(p)
                require(p, character.only = T)
                print(paste0(p, " downloaded and loaded successfully"))
        }
}


## Paths

dat_dir <- "data/"
annot_dir <- "data/annotations/"
fig_dir <- "figures/"
res_dir <- "results/all_platforms/"


## File names

all_platform_data_fname <- here(dat_dir, "all_platforms_data_updated.RData")
res_name_list <- c("_overall", "_female", "_male", "_ics_1y_y", "_ics_1y_n", "_eos_high", "_eos_low", 
                   "_bmi_high", "_bmi_low", "_allergy_y", "_allergy_n", "_neut_high", "_neut_low")
res_asth_fname <- here(res_dir, "res_compare_dif_asthmatics.rds")
res_qt_fname <- here(res_dir, "res_qt_asthmatics.rds")
annot_fname <- here(annot_dir, "PEGASuS_Annotations_PeiZhang_MH.xlsx")
sph_dat_fname <- here(dat_dir, "sph_data_updated.RData")
microb_dat_fname <- here(dat_dir, "microb_data_updated.RData")


## Hard-coded numbers

missing_thld <- 0.30
sig_digits <- 3
alpha_thld <- 0.05
z_95 <- qnorm(p = 0.975)
keep_cols <- c("subgroup", "feat_id", "tentative_annotation", "med_feat", "wo_ms2", "feature", "platform", 
               "totaln", "casen", "beta", "pval", "fdr_bh", "or", "lower95", "upper95", "pct_na_met", "cv", 
               "expo_linreg", "beta_linreg", "pval_linreg")
num_cols <- c("beta", "pval", "fdr_bh", "or", "lower95", "upper95", "pct_na_met", "cv", "beta_linreg", 
              "pval_linreg", "neg_log10pval")

```

# 1. Load data & analyses results
## 1.1. Asthmatics vs controls results 

```{r}

load(all_platform_data_fname, verbose = T)

res_all <- NULL
for (n in res_name_list) {
        eval(parse(text = str_c("res", n, " <- readRDS(file = here(res_dir, 'res", n, ".rds'))")))
        eval(parse(text = str_c("res_all <- rbind(res_all, res", n, "[, ..keep_cols])")))
        eval(parse(text = str_c("rm(res", n, ")")))
}


## Sample sizes

unique(res_all[platform == "HILIC-pos", .(subgroup, totaln, casen)]) %>% 
        datatable(caption = "Sample sizes")


## Number of nominally significant features

nrow(res_all[subgroup == "overall" & tentative_annotation != "Unknown" & pval < alpha_thld, ])
nrow(res_all[subgroup == "overall" & tentative_annotation == "Unknown" & pval < alpha_thld, ])


## Add updated annotations from Pei with my tentative pathway annotations

annot_hilic_pos <- read_xlsx(annot_fname, sheet = "HILIC_POS") %>% as.data.table()
annot_hilic_pos[, feat_id := str_c(feature, "_hilic_pos")]
annot_lipid_pos <- read_xlsx(annot_fname, sheet = "LIPIDOMICS_POS") %>% as.data.table()
annot_lipid_pos[, feat_id := str_c(feature, "_lipid_pos")]
annot_lipid_neg <- read_xlsx(annot_fname, sheet = "LIPIDOMICS_NEG") %>% as.data.table()
annot_lipid_neg[, feat_id := str_c(feature, "_lipid_neg")]
annot <- rbind(annot_hilic_pos, annot_lipid_pos, annot_lipid_neg)

annot <- annot[feat_id != "F324_lipid_neg", ] # Remove duplicated cortisol (lipid_neg)

annot[, .N, confidence_level]
annot[, annotation := case_when(confidence_level == "AMRT" ~ annotation, 
                                confidence_level %in% c("AMMS2", "MS2") ~ str_c(annotation, "*"), 
                                confidence_level == "AM" ~ str_c(annotation, "**"))]

datatable(annot[, .N, .(tentative_subpathway,	tentative_superpathway)][order(-N)], 
          caption = "Features with tentatively annotated pathways", filter = "top")


## Add in targeted sphingolipids metabolites annotations

sph_dat <- load(sph_dat_fname)
sph_mets_info[, ':='(annotation = metabolite, 
                     confidence_level = NA, 
                     comment = NA, 
                     tentative_subpathway = "Sphingolipids (Targeted)",	
                     tentative_superpathway = "Sphingolipids (T)",	
                     pathway_comment = NA)]


## Add in targeted microbial metabolites annotations

microb_dat <- load(microb_dat_fname)
microb_mets_info[, ':='(annotation = Component_Name, 
                        confidence_level = NA, 
                        comment = NA, 
                        tentative_subpathway = gsub("polyamine metabolism", "Polyamine metabolism", SUB_PATHWAY),	
                        tentative_superpathway = "Microbial Metabolites (T)",	
                        pathway_comment = NA)]

annot_all <- rbind(annot[, .(feat_id, annotation, confidence_level, comment, 
                             tentative_subpathway,	tentative_superpathway,	pathway_comment)], 
                   sph_mets_info[, .(feat_id, annotation, confidence_level, comment, 
                                     tentative_subpathway,	tentative_superpathway,	pathway_comment)], 
                   microb_mets_info[, .(feat_id, annotation, confidence_level, comment, 
                                        tentative_subpathway,	tentative_superpathway,	pathway_comment)])
annot_all[, tentative_superpathway := factor(tentative_superpathway, 
                                             levels = c("Amino Acid", "Carbohydrate", "Cofactors and Vitamins", 
                                                        "Lipid", "Nucleotide", "Peptide", "Xenobiotics", 
                                                        "Microbial Metabolites (T)", "Sphingolipids (T)", "Unknown"))]
annot_all <- annot_all[order(tentative_superpathway, tentative_subpathway, annotation)]
annot_all[, ':='(tentative_subpathway = factor(tentative_subpathway, levels = unique(annot_all$tentative_subpathway)), 
                 feat_id_num = seq(1, .N))]
res_all <- merge(res_all, annot_all, by = "feat_id", all.x = T)


## Write out mets info for all five platforms

fwrite(hilic_pos_mets_info, file = here(dat_dir, "hilic_pos_feature_info_qc.csv"))
fwrite(lipid_pos_mets_info, file = here(dat_dir, "lipid_pos_feature_info_qc.csv"))
fwrite(lipid_neg_mets_info, file = here(dat_dir, "lipid_neg_feature_info_qc.csv"))
fwrite(sph_mets_info, file = here(dat_dir, "sphingolipid_metabolites_info_qc.csv"))
fwrite(microb_mets_info, file = here(dat_dir, "microbial_metabolites_info_qc.csv"))

```

## 1.2. Within asthamtics results

```{r}

res_asth <- readRDS(res_asth_fname)
res_asth <- merge(res_asth, annot[, .(feat_id, annotation, confidence_level, comment, 
                                      tentative_subpathway,	tentative_superpathway,	pathway_comment)], 
                  by = "feat_id", all.x = T)

res_qt <- readRDS(res_qt_fname)
res_qt <- merge(res_qt, annot[, .(feat_id, annotation, confidence_level, comment, 
                                  tentative_subpathway,	tentative_superpathway,	pathway_comment)], 
                by = "feat_id", all.x = T)

```

## 1.3. Concatenate all results

```{r}

res_all[grepl("Montelukast|Mequitazine|Acetaminophen|Omeprazole|Tolmetin", annotation, ignore.case = T) | 
          tentative_subpathway == "Drug metabolism", med_feat := 1][is.na(med_feat), med_feat := 0]
res_all[, ':='(feat_id_plot = ifelse(is.na(annotation), str_c(feature, "_", tentative_annotation), 
                                     str_c(feature, "_", annotation)), 
               med_feat = as.factor(med_feat), 
               subgroup = factor(subgroup, levels = c("overall", "female", "male", "ics_1y_prior_yes", 
                                                      "ics_1y_prior_no", "eos_high_300", "eos_low_300", 
                                                      "bmi_high_30", "bmi_low_30", "ige_high_env_allergy", 
                                                      "ige_norm_no_env_allergy", "neut_high_median", "neut_low_median")), 
               platform = as.factor(platform), 
               neg_log10pval = -log10(pval), 
               direction_of_effect = case_when(beta <= 0 ~ "negative", 
                                               beta > 0 ~ "positive"))]
setcolorder(res_all, c("platform", "feat_id", "annotation", "confidence_level", "comment", "subgroup", 
                       "totaln", "beta", "pval", "fdr_bh", "tentative_subpathway", "tentative_superpathway"))


## Indicated untargeted metabolites already in the targeted platform

res_all[feat_id_plot %in% c("F247_Sphingosine**", "F1461_Sphingosine 1-phosphate**", "F1083_Sphingosine**", 
                            "F1216_D-erythro-Sphingosine C-20**", "F1467_D-erythro-Sphinganine-1-phosphate**", 
                            "F475_Serotonin*", "F544_1H-Indole-3-propanoic acid**"), 
        repeated_in_targeted := 1]
saveRDS(res_all, file = here(res_dir, "res_all.rds"))


## Features with p<0.05 in overall analyses

res_all[subgroup == "overall" & pval < alpha_thld, .N, platform][order(-N)]

# res_all[subgroup == "overall" & pval < alpha_thld & !is.na(annotation), ][order(pval)] %>%
#         datatable(filter = "top") %>% formatSignif(columns = num_cols, digits = 3)
feat_overall_005 <- res_all[subgroup == "overall" & pval < alpha_thld & !is.na(annotation), feat_id]
length(feat_overall_005)


## Features with p<0.01 in overall analyses

feat_overall_001 <- res_all[subgroup == "overall" & pval < alpha_thld/5 & !is.na(annotation), feat_id]
length(feat_overall_001)

# fwrite(res_all[feat_id %in% feat_overall_005 & subgroup == "overall", ], 
#        file = here(res_dir, "features_significant_in_overall_nominal.csv"))


## Top features in subgroup analyses

feat_eoshigh_001 <- res_all[subgroup == "eos_high_300" & pval < alpha_thld/5 
                            & tentative_annotation != "Unknown", unique(feat_id)]
feat_bmihigh_001 <- res_all[subgroup == "bmi_high_30" & pval < alpha_thld/5 
                            & tentative_annotation != "Unknown", unique(feat_id)]
feat_allergy_001 <- res_all[subgroup == "ige_high_env_allergy" & pval < alpha_thld/5 
                            & tentative_annotation != "Unknown", unique(feat_id)]
feat_all_001 <- res_all[pval < alpha_thld/5 & tentative_annotation != "Unknown", unique(feat_id)]
feat_all_fdr005 <- res_all[fdr_bh < alpha_thld & tentative_annotation != "Unknown", unique(feat_id)]

feat_eoshigh_001 %in% annot_all$feat_id %>% table()
feat_bmihigh_001 %in% annot_all$feat_id %>% table()
feat_allergy_001 %in% annot_all$feat_id %>% table()
feat_all_001 %in% annot_all$feat_id %>% table()
feat_all_fdr005 %in% annot_all$feat_id %>% table()

# View(mets_info_all[feat_id %in% feat_all_001, ][!(feat_id %in% annot_all$feat_id), ])
# To look at tentative annotation of those not included in the re-annotated features

```

## 1.4. Features with p<0.05 in overall analysis 

> include features fdr<0.05 in subgroup analyses  

Subgroup analyses:  
 - by sex F/M  
 - by number of ICS prescriptions within 1 year prior to serum collection > 0 or not  
 - by eosinophil count closest to serum collection >= 300/uL or not  
 - by number of asthma exacerbations within 1 year prior to serum collection > 0 or not  
 - by BMI >= 30 or not  
 - by total/total IgE high or existence of environmental allergy 
 - by neutrophil count higher than median or not  

### 1.4.1. Results tables

```{r}

datatable(res_all[feat_id %in% feat_overall_005, ][order(subgroup, pval)], filter = "top", 
          caption = "Features with p<0.05 in overall analysis from global platforms") %>% 
        formatSignif(columns = num_cols, digits = 3)

datatable(res_all[feat_id %in% feat_overall_005 & grepl("Targeted", platform), ][order(subgroup, pval)], filter = "top", 
          caption = "Features with p<0.05 in overall analysis from targeted platforms") %>% 
        formatSignif(columns = num_cols, digits = 3)

datatable(res_all[subgroup == "eos_high_300" & pval < alpha_thld & tentative_annotation != "Unknown"
                  ][order(subgroup, pval)], 
          filter = "top", caption = "Features with p<0.05 in eosinophilic asthma vs controls") %>% 
        formatSignif(columns = num_cols, digits = 3)

datatable(res_all[subgroup == "bmi_high_30" & pval < alpha_thld & tentative_annotation != "Unknown"
                  ][order(subgroup, pval)], 
          filter = "top", caption = "Features with p<0.05 in obese asthma vs controls") %>% 
        formatSignif(columns = num_cols, digits = 3)

datatable(res_all[subgroup == "ige_high_env_allergy" & pval < alpha_thld & tentative_annotation != "Unknown"
                  ][order(subgroup, pval)], 
          filter = "top", caption = "Features with p<0.05 in allergic asthma vs controls") %>% 
        formatSignif(columns = num_cols, digits = 3)

```

### 1.4.2. Heatmaps of negative log10(p-value) (global platform features)

```{r}

save_pheatmap_png <- function(x, filename, width = 18, height = 18, units = "in", res = 300) {
        png(filename, width = width, height = height, units = units, res = res)
        grid::grid.newpage()
        grid::grid.draw(x$gtable)
        dev.off()
}


## p < 0.01 

res_plot_001 <- res_all[feat_id %in% feat_overall_001 & !grepl("Targeted", platform), 
                        ][order(subgroup, tentative_superpathway, tentative_subpathway, feat_id_plot)]
res_plot_001[, feat_id_plot := factor(feat_id_plot, levels = res_plot_001[subgroup == "overall", feat_id_plot])]

pval_mat_001 <- dcast(res_plot_001, feat_id_plot ~ subgroup, value.var = "neg_log10pval") %>% as.data.frame()
rownames(pval_mat_001) <- pval_mat_001$feat_id_plot
pval_mat_001$feat_id_plot <- NULL

subway_annot_001 <- res_plot_001[subgroup == "overall", .(feat_id_plot, tentative_subpathway, tentative_superpathway)
                                 ][order(tentative_superpathway, tentative_subpathway, feat_id_plot)] %>% as.data.frame()
rownames(subway_annot_001) <- subway_annot_001$feat_id_plot
subway_annot_001$feat_id_plot <- NULL

pheatmap(pval_mat_001, cluster_rows = F, cluster_cols = F, angle_col = 315, fontsize = 16, 
         annotation_row = subway_annot_001, main = "-log10(p-value)", color = inferno(10, direction = -1)) %>% 
        save_pheatmap_png(., here(fig_dir, "heatmaps/heatmap_neg_log10pval_001.png"), width = 21, height = 24)
pheatmap(subset(pval_mat_001, select = c("overall", "eos_high_300", "bmi_high_30", "ige_high_env_allergy")), 
         cluster_rows = F, cluster_cols = F, angle_col = 315, fontsize = 16, annotation_row = subway_annot_001, 
         main = "-log10(p-value)", color = inferno(10, direction = -1)) %>% 
        save_pheatmap_png(., here(fig_dir, "heatmaps/heatmap_neg_log10pval_001_subtype.png"), width = 18, height = 24)


## p < 0.05

res_plot_005 <- res_all[feat_id %in% feat_overall_005 & !grepl("Targeted", platform), 
                        ][order(subgroup, tentative_superpathway, tentative_subpathway, feat_id_plot)]
res_plot_005[, feat_id_plot := factor(feat_id_plot, levels = res_plot_005[subgroup == "overall", feat_id_plot])]

pval_mat_005 <- dcast(res_plot_005, feat_id_plot ~ subgroup, value.var = "neg_log10pval") %>% as.data.frame()
rownames(pval_mat_005) <- pval_mat_005$feat_id_plot
pval_mat_005$feat_id_plot <- NULL

subway_annot_005 <- res_plot_005[subgroup == "overall", .(feat_id_plot, tentative_subpathway, tentative_superpathway)
                                 ][order(tentative_superpathway, tentative_subpathway, feat_id_plot)] %>% as.data.frame()
rownames(subway_annot_005) <- subway_annot_005$feat_id_plot
subway_annot_005$feat_id_plot <- NULL

pheatmap(pval_mat_005, cluster_rows = F, cluster_cols = F, angle_col = 315, fontsize = 16, 
         annotation_row = subway_annot_005, main = "-log10(p-value)", color = inferno(10, direction = -1)) %>% 
        save_pheatmap_png(., here(fig_dir, "heatmaps/heatmap_neg_log10pval_005.png"), width = 21, height = 42)
pheatmap(subset(pval_mat_005, select = c("overall", "eos_high_300", "bmi_high_30", "ige_high_env_allergy")), 
         cluster_rows = F, cluster_cols = F, angle_col = 315, fontsize = 16, annotation_row = subway_annot_005, 
         main = "-log10(p-value)", color = inferno(10, direction = -1)) %>% 
        save_pheatmap_png(., here(fig_dir, "heatmaps/heatmap_neg_log10pval_005_subtype.png"), width = 18, height = 42)


# ## beta
# 
# beta_mat <- dcast(res_all[feat_id %in% c(feat_overall_001, feat_subgrp_fdr_005) & med_feat == 0, ], feat_id ~ subgroup, 
#                   value.var = "beta") 
# rownames(beta_mat) <- beta_mat$feat_id
# beta_mat$feat_id <- NULL
# pheatmap(beta_mat, cluster_rows = T, cluster_cols = F, 
#          main = "beta estimate") %>% 
#         save_pheatmap_png(., here(fig_dir, "heatmap_beta.png"))

```

### 1.4.3. Venn Diagram

```{r}

list(`overall` = res_all[subgroup == "overall" & pval < alpha_thld & !grepl("Targeted", platform), 
                         feat_id], 
     `eos_high_300` = res_all[subgroup == "eos_high_300" & pval < alpha_thld & !grepl("Targeted", platform), 
                              feat_id], 
     `ige_high_env_allergy` = res_all[subgroup == "ige_high_env_allergy" & pval < alpha_thld 
                                      & !grepl("Targeted", platform), feat_id], 
     `bmi_high_30` = res_all[subgroup == "bmi_high_30" & pval < alpha_thld 
                             & !grepl("Targeted", platform), feat_id]) %>%
        ggvenn(stroke_size = 0.5, set_name_size = 5, text_size = 5, show_percentage = F) +  
               labs(title = "Number of features associated with asthma status (p<0.05)") + 
               theme(plot.title = element_text(hjust = 0.5, size = 15))
ggsave(file = here(fig_dir, "venn_diagrams/venn_overall_subtype_pval005.png"), width = 8, height = 8)

list(`overall` = res_all[subgroup == "overall" & pval < alpha_thld/5 & !grepl("Targeted", platform), 
                         feat_id], 
     `eos_high_300` = res_all[subgroup == "eos_high_300" & pval < alpha_thld/5 & !grepl("Targeted", platform), 
                              feat_id], 
     `ige_high_env_allergy` = res_all[subgroup == "ige_high_env_allergy" & pval < alpha_thld/5 
                                      & !grepl("Targeted", platform), feat_id], 
     `bmi_high_30` = res_all[subgroup == "bmi_high_30" & pval < alpha_thld/5 
                             & !grepl("Targeted", platform), feat_id]) %>%
        ggvenn(stroke_size = 0.5, set_name_size = 5, text_size = 5, show_percentage = F) +  
               labs(title = "Number of features associated with asthma status (p<0.01)") + 
               theme(plot.title = element_text(hjust = 0.5, size = 15))
ggsave(file = here(fig_dir, "venn_diagrams/venn_overall_subtype_pval001.png"), width = 8, height = 8)

list(`overall` = res_all[subgroup == "overall" & fdr_bh < alpha_thld & !grepl("Targeted", platform), 
                         feat_id], 
     `eos_high_300` = res_all[subgroup == "eos_high_300" & fdr_bh < alpha_thld & !grepl("Targeted", platform), 
                              feat_id], 
     `ige_high_env_allergy` = res_all[subgroup == "ige_high_env_allergy" & fdr_bh < alpha_thld 
                                      & !grepl("Targeted", platform), feat_id], 
     `bmi_high_30` = res_all[subgroup == "bmi_high_30" & fdr_bh < alpha_thld 
                             & !grepl("Targeted", platform), feat_id]) %>%
        ggvenn(stroke_size = 0.5, set_name_size = 5, text_size = 5, show_percentage = F) +  
               labs(title = "Number of features associated with asthma status (fdr<0.05)") + 
               theme(plot.title = element_text(hjust = 0.5, size = 15))
ggsave(file = here(fig_dir, "venn_diagrams/venn_overall_subtype_fdr005.png"), width = 8, height = 8)

```

### 1.4.4. Manhanttan plots

```{r}

save_manh_plot <- function(dat, title_infig, fig_name, wdth = 20, ht = 12, pointsize = 3, pointstroke = 1.5,
                           transparency = 1, if_lbl = c(T, F), lbl_thld = 1e-6) {
  
        p <- ggplot(dat, aes(x = feat_id_num, y = neg_log10pval)) + 
                    facet_grid(. ~ tentative_superpathway, scales = "free_x") + # , space = "free_x"
                    geom_point(aes(color = tentative_subpathway, shape = direction_of_effect), 
                               size = pointsize, stroke = pointstroke, alpha = transparency) + 
                    scale_shape_manual(values = c("negative" = 6, "positive" = 2)) + 
                    labs(title = title_infig, 
                         y = expression('-log'[10]*'('*italic(P)*'-value)')) + 
                    theme(legend.position = "bottom", 
                          legend.text = element_text(size = 10), 
                          plot.title = element_text(hjust = 0.5, size = 16), 
                          strip.text.x = element_text(size = 15), 
                          axis.title.x = element_blank(),
                          axis.text.x = element_blank(), 
                          axis.ticks.x = element_blank(), 
                          axis.title.y = element_text(size = 15), 
                          axis.text.y = element_text(size = 12))
        if (if_lbl == T) {
                p <- p + geom_label_repel(aes(label = ifelse(pval < lbl_thld, annotation, "")),
                                          box.padding   = 0.25,
                                          point.padding = 0.5,
                                          segment.color = 'grey50',
                                          max.overlaps = 80)
        }
        ggsave(file = here(fig_dir, fig_name), plot = p, width = wdth, height = ht)
}

save_manh_plot(dat = res_plot_005[subgroup == "overall", ], 
               title_infig = "Asthmatics vs controls overall analysis (P<0.05)", 
               fig_name = "manhattan_plots/manhattan_asth_vs_ctrl_overall_p005.png", 
               if_lbl = F, wdth = 27)
save_manh_plot(dat = res_plot_001[subgroup == "overall", ], 
               title_infig = "Asthmatics vs controls overall analysis (P<0.01)", 
               fig_name = "manhattan_plots/manhattan_asth_vs_ctrl_overall_p001.png", 
               if_lbl = T, lbl_thld = 1e-4, wdth = 36)

save_manh_plot(dat = res_plot_005[subgroup == "eos_high_300", ], 
               title_infig = "Asthmatics with eosinophil >= 300U/L vs controls (P<0.05)", 
               fig_name = "manhattan_plots/manhattan_asth_vs_ctrl_eos_high_300_p005.png", 
               if_lbl = T, lbl_thld = 1e-2, wdth = 27)
save_manh_plot(dat = res_plot_001[subgroup == "eos_high_300", ], 
               title_infig = "Asthmatics with eosinophil >= 300U/L vs controls (P<0.01)", 
               fig_name = "manhattan_plots/manhattan_asth_vs_ctrl_eos_high_300_p001.png", 
               if_lbl = T, lbl_thld = 1e-2, wdth = 27)

save_manh_plot(dat = res_plot_005[subgroup == "bmi_high_30", ], 
               title_infig = "Asthmatics with BMI >= 30kg/m2 vs controls (P<0.05)", 
               fig_name = "manhattan_plots/manhattan_asth_vs_ctrl_bmi_high_30_p005.png", 
               if_lbl = T, lbl_thld = 1e-2, wdth = 27)
save_manh_plot(dat = res_plot_001[subgroup == "bmi_high_30", ], 
               title_infig = "Asthmatics with BMI >= 30kg/m2 vs controls (P<0.01)", 
               fig_name = "manhattan_plots/manhattan_asth_vs_ctrl_bmi_high_30_p001.png", 
               if_lbl = T, lbl_thld = 1e-2, wdth = 27)

save_manh_plot(dat = res_plot_005[subgroup == "ige_high_env_allergy", ], 
               title_infig = "Asthmatics with high IgE or environmental allergy vs controls (P<0.05)", 
               fig_name = "manhattan_plots/manhattan_asth_vs_ctrl_ige_high_env_allergy_p005.png", 
               if_lbl = T, lbl_thld = 1e-2, wdth = 27)
save_manh_plot(dat = res_plot_001[subgroup == "ige_high_env_allergy", ], 
               title_infig = "Asthmatics with high IgE or environmental allergy vs controls (P<0.01)", 
               fig_name = "manhattan_plots/manhattan_asth_vs_ctrl_ige_high_env_allergy_p001.png", 
               if_lbl = T, lbl_thld = 1e-2, wdth = 27)

```

### 1.4.5. Effect estimate plots

```{r}

save_eff_est_plot <- function(dat, title_infig, xaxis_name = "Serum Metabolite Feature", yaxis_name, 
                              fig_name, wdth = 8, ht = 12) {

        p <- ggplot(dat = dat, aes(x = feat_id_plot, y = or)) + 
                    facet_grid(. ~ subgroup) +
                    geom_errorbar(aes(ymin = lower95, ymax = upper95, color = tentative_subpathway), width = 0.25) + 
                    geom_pointrange(aes(ymin = lower95, ymax = upper95, color = tentative_subpathway, 
                                        shape = pval < 0.05)) + 
                    # scale_colour_manual(values = c("negative" = "#2578A1", "positive" = "#DB9B3C")) + 
                    scale_shape_manual(values = c("TRUE" = 8, "FALSE" = 1)) +
                    geom_hline(yintercept = 1, linetype = 2) + 
                    labs(title = title_infig, x = xaxis_name, y = yaxis_name) + 
                    # theme_minimal() + 
                    theme(title = element_text(size = 16), 
                          axis.title = element_text(size = 16), 
                          axis.text = element_text(size = 12), 
                          strip.text.x = element_text(size = 16), 
                          legend.position = "right",
                          legend.title = element_text(size = 12), 
                          legend.text = element_text(size = 12)) + 
                    scale_y_continuous(trans = log_trans(), 
                                       breaks = axisTicks(log(range(dat[, or])), 
                                                          log = TRUE)) + 
                    coord_flip() 
        ggsave(here(fig_dir, fig_name), plot = p, width = wdth, height = ht)
}

save_eff_est_plot(dat = res_plot_005[subgroup %in% c("overall", "eos_high_300", "bmi_high_30", "ige_high_env_allergy") 
                                     & grepl("(?i)ster", tentative_subpathway) & is.na(repeated_in_targeted), ], 
                  title_infig = "Asthmatics vs controls analysis (P<0.05)", 
                  xaxis_name = "Steroid metabolites", 
                  yaxis_name = "Odds ratio (95% confidence interval)", 
                  fig_name = "effect_estimate_plots/eff_est_asth_vs_ctrl_steroids_p005.png", 
                  wdth = 20, ht = 10) 
save_eff_est_plot(dat = res_plot_001[subgroup %in% c("overall", "eos_high_300", "bmi_high_30", "ige_high_env_allergy") 
                                     & grepl("(?i)ster", tentative_subpathway) & is.na(repeated_in_targeted), ], 
                  title_infig = "Asthmatics vs controls analysis (P<0.01)", 
                  xaxis_name = "Steroid metabolites", 
                  yaxis_name = "Odds ratio (95% confidence interval)", 
                  fig_name = "effect_estimate_plots/eff_est_asth_vs_ctrl_steroids_p001.png", 
                  wdth = 20, ht = 10) 

save_eff_est_plot(dat = res_plot_005[subgroup %in% c("overall", "eos_high_300", "bmi_high_30", "ige_high_env_allergy") 
                                     & grepl("(?i)ceramide|sphing", tentative_subpathway) & is.na(repeated_in_targeted), ], 
                  title_infig = "Asthmatics vs controls analysis (P<0.05)", 
                  xaxis_name = "Sphingolipid metabolites", 
                  yaxis_name = "Odds ratio (95% confidence interval)", 
                  fig_name = "effect_estimate_plots/eff_est_asth_vs_ctrl_sphingolipids_p005.png", 
                  wdth = 20, ht = 10) 
save_eff_est_plot(dat = res_plot_001[subgroup %in% c("overall", "eos_high_300", "bmi_high_30", "ige_high_env_allergy") 
                                     & grepl("(?i)ceramide|sphing", tentative_subpathway) & is.na(repeated_in_targeted), ], 
                  title_infig = "Asthmatics vs controls analysis (P<0.01)", 
                  xaxis_name = "Sphingolipid metabolites", 
                  yaxis_name = "Odds ratio (95% confidence interval)", 
                  fig_name = "effect_estimate_plots/eff_est_asth_vs_ctrl_sphingolipids_p001.png", 
                  wdth = 20, ht = 4) 

save_eff_est_plot(dat = res_plot_005[subgroup %in% c("overall", "eos_high_300", "bmi_high_30", "ige_high_env_allergy") 
                                     & grepl("(?i)tryptophan|arginine|cysteine|taurine", tentative_subpathway)
                                     & is.na(repeated_in_targeted), ], 
                  title_infig = "Asthmatics vs controls analysis (P<0.05)", 
                  xaxis_name = "Amino acid metabolites", 
                  yaxis_name = "Odds ratio (95% confidence interval)", 
                  fig_name = "effect_estimate_plots/eff_est_asth_vs_ctrl_aminoacid_p005.png", 
                  wdth = 20, ht = 10) 
save_eff_est_plot(dat = res_plot_001[subgroup %in% c("overall", "eos_high_300", "bmi_high_30", "ige_high_env_allergy") 
                                     & grepl("(?i)tryptophan|arginine|cysteine|taurine", tentative_subpathway)
                                     & is.na(repeated_in_targeted), ], 
                  title_infig = "Asthmatics vs controls analysis (P<0.01)", 
                  xaxis_name = "Amino acid metabolites", 
                  yaxis_name = "Odds ratio (95% confidence interval)", 
                  fig_name = "effect_estimate_plots/eff_est_asth_vs_ctrl_aminoacid_p001.png", 
                  wdth = 20, ht = 10) 

save_eff_est_plot(dat = res_plot_005[subgroup %in% c("overall", "eos_high_300", "bmi_high_30", "ige_high_env_allergy") 
                                     & grepl("(?i)lyso|phospholipid", tentative_subpathway)
                                     & is.na(repeated_in_targeted), ], 
                  title_infig = "Asthmatics vs controls analysis (P<0.05)", 
                  xaxis_name = "Glycerophospholipid metabolites", 
                  yaxis_name = "Odds ratio (95% confidence interval)", 
                  fig_name = "effect_estimate_plots/eff_est_asth_vs_ctrl_glycerophospholipid_p005.png", 
                  wdth = 20, ht = 20) 
save_eff_est_plot(dat = res_plot_001[subgroup %in% c("overall", "eos_high_300", "bmi_high_30", "ige_high_env_allergy") 
                                     & grepl("(?i)lyso|phospholipid", tentative_subpathway)
                                     & is.na(repeated_in_targeted), ], 
                  title_infig = "Asthmatics vs controls analysis (P<0.01)", 
                  xaxis_name = "Glycerophospholipid metabolites", 
                  yaxis_name = "Odds ratio (95% confidence interval)", 
                  fig_name = "effect_estimate_plots/eff_est_asth_vs_ctrl_glycerophospholipid_p001.png", 
                  wdth = 20, ht = 15) 

```


> Lipidomics nomenclature

 | Abbreviation | Lipid subclass | Main class | 
 | ------------ | -------------- | ---------- | 
 | Cer | Ceramide | Ceramides | 
 | CerP | Ceramide 1-phosphates | Ceramides | 
 | FAHFA | Fatty acid ester of hydroxyl fatty acid | Fatty esters | 
 | HexCer | Hexosylceramide | Neutral glycosphingolipids | 
 | Hex2Cer | Dihexosylceramide | Neutral glycosphingolipids | 
 | Hex3Cer | Trihexosylceramide | Neutral glycosphingolipids | 
 | LPC | Lysophophatidylcholine | Glycerophosphocholines | 
 | LPE | Lysophosphatidylethanolamine | Glycerophosphoethanolamines | 
 | LPI | Lysophosphatidylinositol | Glycerophosphoinositols | 
 | NAE | N-acyl ethanolamines | Fatty amides | 
 | TG | Triacylglycerol | Triradylglycerols | 
 | PA | Phosphatidic acid | Glycerophosphates | 
 | PC | Phosphatidylcholine | Glycerophosphocholines | 
 | PE | Phosphatidylethanolamine | Glycerophosphoethanolamines | 
 | PMeOH | Phosphatidylmethanol | Other Glycerophospholipids | 


# 2. Within asthmatics comparison results
## 2.1. Binary comparisons

```{r, eval = F}

res_asth[grepl("Montelukast|Mequitazine|Acetaminophen|Omeprazole|Tolmetin", annotation, ignore.case = T) | 
           tentative_subpathway == "Drug metabolism", med_feat := 1][is.na(med_feat), med_feat := 0]
res_asth[, fdr_bh := p.adjust(pval, method = "BH"), by = comparison]
res_asth[, ':='(feat_id_plot = ifelse(is.na(annotation), str_c(feature, "_", tentative_annotation), 
                                      str_c(feature, "_", annotation)), 
                med_feat = as.factor(med_feat), 
                comparison = factor(comparison, 
                                    levels = c("high_vs_low_eos_asthmatics", "obese_vs_not_asthmatics", 
                                               "allergy_vs_not_asthmatics", "female_vs_male_asthmatics", 
                                               "female_vs_male_controls")), 
                platform = as.factor(platform), 
                neg_log10pval = -log10(pval), 
                direction_of_effect = case_when(beta <= 0 ~ "negative", 
                                                beta > 0 ~ "positive"))]
setcolorder(res_asth, c("platform", "feat_id", "annotation", "confidence_level", "comment", "comparison", 
                        "beta", "pval", "fdr_bh"))
saveRDS(res_asth, file = here(res_dir, "res_asth.rds"))

datatable(res_asth[pval < alpha_thld & !(is.na(annotation) & tentative_annotation == "Unknown"), c(1:18)
                   ][order(comparison, pval)], filter = "top") %>% 
        formatSignif(columns = c("beta", "pval", "fdr_bh", "cv", "pct_na_met"), digits = 3)


## -log10(pval)

pval_mat_asth_001 <- dcast(res_asth[feat_id %in% feat_overall_001 & med_feat == 0, ], feat_id_plot ~ comparison, 
                           value.var = "neg_log10pval") %>% as.data.frame()
rownames(pval_mat_asth_001) <- pval_mat_asth_001$feat_id_plot
pval_mat_asth_001$feat_id_plot <- NULL
pheatmap(pval_mat_asth_001, cluster_rows = F, cluster_cols = F, angle_col = 315, fontsize = 16, 
         main = "-log10(p-value)", color = inferno(10, direction = -1)) %>% 
        save_pheatmap_png(., here(fig_dir, "heatmaps/heatmap_neg_log10pval_001_compare_asthma_subtypes.png"), 
                          width = 12, height = 21)


pval_mat_asth_005 <- dcast(res_asth[feat_id %in% feat_overall_005 & med_feat == 0, ], feat_id_plot ~ comparison, 
                           value.var = "neg_log10pval") %>% as.data.frame()
rownames(pval_mat_asth_005) <- pval_mat_asth_005$feat_id_plot
pval_mat_asth_005$feat_id_plot <- NULL
pheatmap(pval_mat_asth_005, cluster_rows = F, cluster_cols = F, angle_col = 315, fontsize = 16, 
         main = "-log10(p-value)", color = inferno(20, direction = -1)) %>% 
        save_pheatmap_png(., here(fig_dir, "heatmaps/heatmap_neg_log10pval_005_compare_asthma_subtypes.png"), 
                          width = 12, height = 42)


## Venn Diagram

list(`high_vs_low_eos` = res_asth[comparison == "high_vs_low_eos_asthmatics" & pval < alpha_thld, feat_id], 
     `obese_vs_not` = res_asth[comparison == "obese_vs_not_asthmatics" & pval < alpha_thld, feat_id], 
     `allergy_vs_not` = res_asth[comparison == "allergy_vs_not_asthmatics" & pval < alpha_thld, feat_id], 
     `female_vs_male` = res_asth[comparison == "female_vs_male_asthmatics" & pval < alpha_thld, feat_id]) %>%
        ggvenn(stroke_size = 0.5, set_name_size = 5, text_size = 5, show_percentage = F) +  
               labs(title = "Number of features associated with binary outcomes in asthmatics (p<0.05)") + 
               theme(plot.title = element_text(hjust = 0.5, size = 15))
ggsave(file = here(fig_dir, "venn_diagrams/venn_bin_outc_asthmatics_pval005.png"), width = 8, height = 8)

```

## 2.2. Quantitative traits

```{r, eval = F}

res_qt[grepl("Montelukast|Mequitazine|Acetaminophen|Omeprazole|Tolmetin", annotation, ignore.case = T) | 
         tentative_subpathway == "Drug metabolism", med_feat := 1][is.na(med_feat), med_feat := 0]
res_qt[, fdr_bh := p.adjust(pval, method = "BH"), by = comparison]
res_qt[, ':='(feat_id_plot = ifelse(is.na(annotation), str_c(feature, "_", tentative_annotation), 
                                    str_c(feature, "_", annotation)), 
              med_feat = as.factor(med_feat), 
              comparison = factor(comparison, 
                                  levels = c("eos_continous", "neut_continous", "fev1_continous", "fvc_continous")), 
              platform = as.factor(platform), 
              neg_log10pval = -log10(pval), 
              direction_of_effect = case_when(beta <= 0 ~ "negative", 
                                              beta > 0 ~ "positive"))]
setcolorder(res_qt, c("platform", "feature", "annotation", "confidence_level", "comment", "comparison", 
                      "beta", "pval", "fdr_bh"))
saveRDS(res_qt, file = here(res_dir, "res_qt.rds"))

datatable(res_qt[pval < alpha_thld & !(is.na(annotation) & tentative_annotation == "Unknown"), c(1:18)
                 ][order(comparison, pval)], filter = "top") %>% 
        formatSignif(columns = c("beta", "pval", "fdr_bh", "cv", "pct_na_met"), digits = 3)


## -log10(pval)

pval_mat_qt_001 <- dcast(res_qt[feat_id %in% feat_overall_001 & med_feat == 0, ], feat_id_plot ~ comparison, 
                                value.var = "neg_log10pval") %>% as.data.frame()
rownames(pval_mat_qt_001) <- pval_mat_qt_001$feat_id_plot
pval_mat_qt_001$feat_id_plot <- NULL
pheatmap(pval_mat_qt_001, cluster_rows = F, cluster_cols = F, angle_col = 315, fontsize = 16, 
         main = "-log10(p-value)", color = inferno(6, begin = 1, end = 0.5, direction = 1)) %>% 
  save_pheatmap_png(., here(fig_dir, "heatmaps/heatmap_neg_log10pval_001_qt.png"), width = 9, height = 21)


pval_mat_qt_005 <- dcast(res_qt[feat_id %in% feat_overall_005 & med_feat == 0, ], feat_id_plot ~ comparison, 
                                value.var = "neg_log10pval") %>% as.data.frame()
rownames(pval_mat_qt_005) <- pval_mat_qt_005$feat_id_plot
pval_mat_qt_005$feat_id_plot <- NULL
pheatmap(pval_mat_qt_005, cluster_rows = F, cluster_cols = F, angle_col = 315, fontsize = 16, 
         main = "-log10(p-value)", color = inferno(12, begin = 1, end = 0.3, direction = 1)) %>% 
  save_pheatmap_png(., here(fig_dir, "heatmaps/heatmap_neg_log10pval_005_qt.png"), width = 9, height = 42)


## Venn Diagram

list(`eos_continous` = res_qt[comparison == "eos_continous" & pval < alpha_thld, feat_id], 
     `neut_continous` = res_qt[comparison == "neut_continous" & pval < alpha_thld, feat_id], 
     `fev1_continous` = res_qt[comparison == "fev1_continous" & pval < alpha_thld, feat_id], 
     `fvc_continous` = res_qt[comparison == "fvc_continous" & pval < alpha_thld, feat_id]) %>%
        ggvenn(stroke_size = 0.5, set_name_size = 5, text_size = 5, show_percentage = F) +  
               labs(title = "Number of features associated with quantitative outcomes in asthmatics (p<0.05)") + 
               theme(plot.title = element_text(hjust = 0.5, size = 15))
ggsave(file = here(fig_dir, "venn_diagrams/venn_qt_outc_asthmatics_pval005.png"), width = 8, height = 8)

list(`eos_continous` = res_qt[comparison == "eos_continous" & fdr_bh < alpha_thld*4, feat_id], 
     `neut_continous` = res_qt[comparison == "neut_continous" & fdr_bh < alpha_thld*4, feat_id], 
     `fev1_continous` = res_qt[comparison == "fev1_continous" & fdr_bh < alpha_thld*4, feat_id], 
     `fvc_continous` = res_qt[comparison == "fvc_continous" & fdr_bh < alpha_thld*4, feat_id]) %>%
        ggvenn(stroke_size = 0.5, set_name_size = 5, text_size = 5, show_percentage = F) +  
               labs(title = "Number of features associated with quantitative outcomes in asthmatics (fdr<0.2)") + 
               theme(plot.title = element_text(hjust = 0.5, size = 15))
# ggsave(file = here(fig_dir, "venn_qt_outc_asthmatics_fdr02.png"), width = 8, height = 8)

```

# Session info

```{r}

timestamp()
sessionInfo()

```
